refactor(cmd): extract flags, validation, and client prep from up.go
